requirements.txt
black==23.1.0
click==8.1.3
colorama==0.4.6
flake8==6.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==0.21.1
tomli==2.0.1
